requirements.txt
django>=4.2
djangorestframework
openpyxl
pandas
whitenoise
gunicorn
psycopg2-binary
